# openapi_output.json
{
    "openapi": "3.1.0",
    "info": {
        "title": "Smith 2.0",
        "description": "Agente de Vendas Inteligente com IA",
        "version": "2.0.0"
    },
    "paths": {
...
        "/api/leads/": {
            "post": {
                "tags": [
                    "Leads"
                ],
                "summary": "Create Lead",
                "description": "Cria um novo lead\n\nArgs:\n    lead_data: Dados do lead a criar\n\nReturns:\n    Lead criado com ID",
                "operationId": "create_lead_api_leads__post",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "$ref": "#/components/schemas/LeadCreate"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/LeadResponse"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "get": {
                "tags": [
                    "Leads"
                ],
                "summary": "List Leads",
                "description": "Lista leads com filtros opcionais\n\nArgs:\n    status: Filtrar por status\n    origem: Filtrar por origem\n    temperatura: Filtrar por temperatura\n    limit: N\u00c3\u00bamero m\u00c3\u00a1ximo de resultados\n    offset: Offset para pagina\u00c3\u00a7\u00c3\u00a3o\n\nReturns:\n    Lista de leads",
                "operationId": "list_leads_api_leads__get",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/LeadStatus"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Status"
                        }
                    },
                    {
                        "name": "origem",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/LeadOrigin"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Origem"
                        }
                    },
                    {
                        "name": "temperatura",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "anyOf": [
                                {
                                    "$ref": "#/components/schemas/LeadTemperature"
                                },
                                {
                                    "type": "null"
                                }
                            ],
                            "title": "Temperatura"
                        }
                    },
                    {
                        "name": "limit",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "maximum": 1000,
                            "default": 100,
                            "title": "Limit"
                        }
                    },
                    {
                        "name": "offset",
                        "in": "query",
                        "required": false,
                        "schema": {
                            "type": "integer",
                            "minimum": 0,
                            "default": 0,
                            "title": "Offset"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "array",
                                    "items": {
                                        "$ref": "#/components/schemas/Lead"
                                    },
                                    "title": "Response List Leads Api Leads  Get"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            }
        },
...
        "/webhook/facebook": {
            "get": {
                "tags": [
                    "Facebook"
                ],
                "summary": "Verify Webhook",
                "description": "Verifica\u00c3\u00a7\u00c3\u00a3o do webhook do Facebook\nFacebook envia GET request para verificar o webhook",
                "operationId": "verify_webhook_webhook_facebook_get",
                "parameters": [
                    {
                        "name": "hub.mode",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Hub.Mode"
                        }
                    },
                    {
                        "name": "hub.verify_token",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Hub.Verify Token"
                        }
                    },
                    {
                        "name": "hub.challenge",
                        "in": "query",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "title": "Hub.Challenge"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    },
                    "422": {
                        "description": "Validation Error",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/HTTPValidationError"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Facebook"
                ],
                "summary": "Receive Facebook Lead",
                "description": "Recebe leads do Facebook Lead Ads\nProcessa, qualifica e insere no CRM automaticamente",
                "operationId": "receive_facebook_lead_webhook_facebook_post",
                "responses": {
                    "200": {
                        "description": "Successful Response",
                        "content": {
                            "application/json": {
                                "schema": {}
                            }
                        }
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "FollowUpConfig": {
                "properties": {
                    "tentativas_realizadas": {
                        "type": "integer",
                        "title": "Tentativas Realizadas",
                        "default": 0
                    },
                    "proxima_tentativa": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Proxima Tentativa"
                    },
                    "intervalo_horas": {
                        "items": {
                            "type": "integer"
                        },
                        "type": "array",
                        "title": "Intervalo Horas",
                        "default": [
                            24,
                            72,
                            168
                        ]
                    },
                    "mensagem_template": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Mensagem Template"
                    }
                },
                "type": "object",
                "title": "FollowUpConfig",
                "description": "Configura\u00c3\u00a7\u00c3\u00a3o de follow-ups"
            },
            "HTTPValidationError": {
                "properties": {
                    "detail": {
                        "items": {
                            "$ref": "#/components/schemas/ValidationError"
                        },
                        "type": "array",
                        "title": "Detail"
                    }
                },
                "type": "object",
                "title": "HTTPValidationError"
            },
            "Lead": {
                "properties": {
                    "id": {
                        "type": "string",
                        "title": "Id"
                    },
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "empresa": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Empresa"
                    },
                    "telefone": {
                        "type": "string",
                        "title": "Telefone"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "status": {
                        "$ref": "#/components/schemas/LeadStatus",
                        "default": "novo"
                    },
                    "origem": {
                        "$ref": "#/components/schemas/LeadOrigin"
                    },
                    "temperatura": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/LeadTemperature"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "qualification_data": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/QualificationData"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "lead_score": {
                        "type": "integer",
                        "title": "Lead Score",
                        "default": 0
                    },
                    "roi_analysis": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/ROIAnalysis"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "valor_estimado": {
                        "type": "number",
                        "title": "Valor Estimado",
                        "default": 0.0
                    },
                    "meeting_scheduled_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Meeting Scheduled At"
                    },
                    "meeting_google_event_id": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Meeting Google Event Id"
                    },
                    "followup_config": {
                        "$ref": "#/components/schemas/FollowUpConfig"
                    },
                    "conversation_history": {
                        "items": {
                            "$ref": "#/components/schemas/ConversationMessage"
                        },
                        "type": "array",
                        "title": "Conversation History",
                        "default": []
                    },
                    "ultima_interacao": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ultima Interacao"
                    },
                    "ultima_mensagem_ia": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ultima Mensagem Ia"
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas"
                    },
                    "tags": {
                        "items": {
                            "type": "string"
                        },
                        "type": "array",
                        "title": "Tags",
                        "default": []
                    },
                    "ai_summary": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ai Summary"
                    },
                    "ai_next_action": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ai Next Action"
                    },
                    "requires_human_approval": {
                        "type": "boolean",
                        "title": "Requires Human Approval",
                        "default": false
                    },
                    "created_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Created At"
                    },
                    "updated_at": {
                        "type": "string",
                        "format": "date-time",
                        "title": "Updated At"
                    },
                    "lost_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Lost At"
                    },
                    "won_at": {
                        "anyOf": [
                            {
                                "type": "string",
                                "format": "date-time"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Won At"
                    },
                    "avatar": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Avatar"
                    }
                },
                "type": "object",
                "required": [
                    "id",
                    "nome",
                    "telefone",
                    "origem"
                ],
                "title": "Lead",
                "description": "Modelo completo de Lead"
            },
            "LeadCreate": {
                "properties": {
                    "nome": {
                        "type": "string",
                        "title": "Nome"
                    },
                    "telefone": {
                        "type": "string",
                        "title": "Telefone"
                    },
                    "empresa": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Empresa"
                    },
                    "email": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Email"
                    },
                    "origem": {
                        "$ref": "#/components/schemas/LeadOrigin"
                    },
                    "notas": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Notas"
                    }
                },
                "type": "object",
                "required": [
                    "nome",
                    "telefone",
                    "origem"
                ],
                "title": "LeadCreate",
                "description": "Payload para criar um lead"
            },
            "LeadOrigin": {
                "type": "string",
                "enum": [
                    "instagram",
                    "whatsapp",
                    "site",
                    "indicacao",
                    "facebook_ads",
                    "teste",
                    "outro"
                ],
                "title": "LeadOrigin",
                "description": "Origem do lead"
            },
            "LeadResponse": {
                "properties": {
                    "success": {
                        "type": "boolean",
                        "title": "Success"
                    },
                    "lead": {
                        "anyOf": [
                            {
                                "$ref": "#/components/schemas/Lead"
                            },
                            {
                                "type": "null"
                            }
                        ]
                    },
                    "message": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Message"
                    }
                },
                "type": "object",
                "required": [
                    "success"
                ],
                "title": "LeadResponse",
                "description": "Response padr\u00c3\u00a3o de Lead"
            },
            "LeadStatus": {
                "type": "string",
                "enum": [
                    "novo",
                    "contato_inicial",
                    "qualificando",
                    "qualificado",
                    "agendamento_marcado",
                    "ganho",
                    "perdido"
                ],
                "title": "LeadStatus",
                "description": "Status do lead no funil de vendas"
            },
            "LeadTemperature": {
                "type": "string",
                "enum": [
                    "quente",
                    "morno",
                    "frio"
                ],
                "title": "LeadTemperature",
                "description": "Temperatura do lead (engajamento)"
            },
...
            "QualificationData": {
                "properties": {
                    "faturamento_anual": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Faturamento Anual"
                    },
                    "is_decision_maker": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Is Decision Maker"
                    },
                    "urgency": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Urgency"
                    },
                    "setor": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Setor"
                    },
                    "wants_roi": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Wants Roi"
                    },
                    "wants_meeting": {
                        "anyOf": [
                            {
                                "type": "boolean"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Wants Meeting"
                    },
                    "atendimentos_por_dia": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Atendimentos Por Dia"
                    },
                    "tempo_por_atendimento": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Tempo Por Atendimento"
                    },
                    "ticket_medio": {
                        "anyOf": [
                            {
                                "type": "number"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ticket Medio"
                    },
                    "funcionarios_atendimento": {
                        "anyOf": [
                            {
                                "type": "integer"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Funcionarios Atendimento"
                    },
                    "maior_desafio": {
                        "anyOf": [
                            {
                                "type": "string"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Maior Desafio"
                    },
                    "ferramentas_atuais": {
                        "anyOf": [
                            {
                                "items": {
                                    "type": "string"
                                },
                                "type": "array"
                            },
                            {
                                "type": "null"
                            }
                        ],
                        "title": "Ferramentas Atuais"
                    }
                },
                "type": "object",
                "title": "QualificationData",
                "description": "Dados de qualifica\u00c3\u00a7\u00c3\u00a3o coletados pela IA"
            },
...
            "ValidationError": {
                "properties": {
                    "loc": {
                        "items": {
                            "anyOf": [
                                {
                                    "type": "string"
                                },
                                {
                                    "type": "integer"
                                }
                            ]
                        },
                        "type": "array",
                        "title": "Location"
                    },
                    "msg": {
                        "type": "string",
                        "title": "Message"
                    },
                    "type": {
                        "type": "string",
                        "title": "Error Type"
                    }
                },
                "type": "object",
                "required": [
                    "loc",
                    "msg",
                    "type"
                ],
                "title": "ValidationError"
            },
...
        },
        "securitySchemes": {
            "HTTPBearer": {
                "type": "http",
                "scheme": "bearer"
            }
        }
    }
}